% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_cp.R
\name{lm_cp}
\alias{lm_cp}
\title{Fit a linear model with a change-point in trend but preserving continuity}
\usage{
lm_cp(formula, data, cp_var = NULL, cp_values = NULL, ...)
}
\arguments{
\item{formula}{An object of class formula}

\item{data}{A data frame containing the variables in the model}

\item{cp_var}{Name of term where change-point is assumed to occur, by default this is the first term on the right-hand side of formula}

\item{cp_values}{Values where change-points may occur}

\item{...}{Additional parameters to be passed to lm}
}
\value{
An object of class lmcp (an augmented object of class lm)
}
\description{
Fit a linear model with a change-point in trend but preserving continuity
}
\examples{
fit <- lm_cp(log(PFUnDA) ~ year, data = pfunda_uria)
fit
summary(fit)
}
