% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfunda_uria}
\alias{pfunda_uria}
\title{PFUnDA data}
\format{
## `pfunda_uria`
A data frame containing a time-series of measurements of PFUnDA (Perfluoroundecanoic acid) in Guillemot (Uria algae) eggs at
Stora Karlsö island in the Baltic Sea
\describe{
  \item{year}{Year of collection}
  \item{PFUnDA}{Concentration (ng/g fresh weight)}
}
}
\usage{
pfunda_uria
}
\description{
PFUnDA data
}
\keyword{datasets}
